Servicios REST facturacion, reconstruccion entidad detalle cuenta
Se crean los demas servicios rest del modulo de facturacion, con respecto a las entidades factura,  detalle factura y detalle cuenta. Se agregan sus relaciones en el abmmanagerfacturacion
Se reconstruye la entidad detalle de cuenta cambiando el tipo de dato del saldo detalle de cuenta de biginteger a int.

+++ pclsystemBackE/src/main/java/com/lawyersys/pclsystembacke/entities/DetalleCuenta.java
@@ -56,7 +56,7 @@ public class DetalleCuenta implements Serializable {
     @Basic(optional = false)
     @NotNull
     @Column(name = "saldo_detalle_cta")
-    private BigInteger saldoDetalleCta;
+    private int saldoDetalleCta;
     
     @JoinColumn(name = "cod_cuenta", referencedColumnName = "cod_cuenta", insertable = false, updatable = false)
     @ManyToOne(optional = false)
@@ -69,7 +69,7 @@ public class DetalleCuenta implements Serializable {
         this.detalleCuentaPK = detalleCuentaPK;
     }
 
-    public DetalleCuenta(DetalleCuentaPK detalleCuentaPK, String descripcion, int monto, boolean estado, BigInteger saldoDetalleCta) {
+    public DetalleCuenta(DetalleCuentaPK detalleCuentaPK, String descripcion, int monto, boolean estado, int saldoDetalleCta) {
         this.detalleCuentaPK = detalleCuentaPK;
         this.descripcion = descripcion;
         this.monto = monto;
@@ -113,11 +113,11 @@ public class DetalleCuenta implements Serializable {
         this.estado = estado;
     }
 
-    public BigInteger getSaldoDetalleCta() {
+    public int getSaldoDetalleCta() {
         return saldoDetalleCta;
     }
 
-    public void setSaldoDetalleCta(BigInteger saldoDetalleCta) {
+    public void setSaldoDetalleCta(int saldoDetalleCta) {
         this.saldoDetalleCta = saldoDetalleCta;
     }
